+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { BrowserRouter, Routes, Route } from 'react-router';
 
-import AboutUs from '@/components/page/About-us-page';
-import Basket from '@/components/page/Basket-page';
-import Catalog from '@/components/page/Catalog-page';
-import Login from '@/components/page/Login-page';
-import Main from '@/components/page/Main-page';
-import ProductDetail from '@/components/page/Product-detail-page';
-import Registration from '@/components/page/Registration-page';
-import UserProfile from '@/components/page/User-profile-page';
+import AboutUs from '@/components/page/about-us-page';
+import Basket from '@/components/page/basket-page';
+import Catalog from '@/components/page/catalog-page';
+import Login from '@/components/page/login-page';
+import Main from '@/components/page/main-page';
+import ProductDetail from '@/components/page/product-detail-page';
+import Registration from '@/components/page/registration-page';
+import UserProfile from '@/components/page/user-profile-page';
 import Navbar from '@/components/ui/navBar/NavBar';
 
 function App() {
