+++ src/components/page/registration-page.tsx
